- compiled with maximum warning level in debug - \Za \Wall \Wx
Former-commit-id: 740f94dd9093fe58fb6a8e4196307b58c33f96a4
Former-commit-id: a9cb603d86765b119aa244435507a0a55c550a15

+++ src/rendering/textureContainer.h
@@ -39,7 +39,7 @@ namespace phi
     private:
         textureContainerLayout _layout;
         size_t _maxTextures;
-        uint _freeSpace;
+        size_t _freeSpace;
         GLint _unit;
         bool _bindless;
         bool _sparse;
